Fix clippy: two errors, dead code, and three lints declared with a reason
`cargo clippy --all-targets` did not just warn, it FAILED. I had never run it
on this repo. Two examples were rejected outright:

    generate_associative.rs:128  (b as u16) >= KEY_LO && ...
    eval_associative.rs:117      b >= KEY_LO && ...

KEY_LO is 0, so the comparison is always true for an unsigned type and
absurd_extreme_comparisons is deny-by-default. `cargo build` passed, so this
only showed up to someone running clippy. Both now say what they mean:
`(KEY_LO..KEY_HI).contains(&b)`. Two more deny-level errors in tests.rs came
from `w.data[0 * 8 + c]`, written that way to mirror `2 * 8 + c` and `3 * 8 +
c`; a `tap(row, col)` accessor keeps the shape visible and drops the
always-zero multiply.

DEAD CODE: `Skeleton.at` and `Skeleton.in_slot` were initialised to 0 and
never read or written again; `as_text` was an unused test helper; VAL_LO and
VAL_HI were unused consts in query_swap.

REAL IMPROVEMENTS, not silencing: a `Table` alias for the n-gram maps, a
slice copy instead of a 4-iteration loop, `contains_key` instead of
`get().is_none()`, `Arch::parse` and `Ptr::offset_by` renamed off the
std-trait names they shadowed, and seven loops whose index really did walk a
single array rewritten as iterators or `fill`.

THREE LINTS SWITCHED OFF for the package, in Cargo.toml, each with its reason
written next to it rather than as scattered #[allow]:
needless_range_loop (the index walks several tensors at different offsets --
`cur[c]`, `aa.data[c]`, `kk.data[t*d+c]` -- so iterating one leaves the rest
indexed anyway), too_many_arguments (kernel signatures), and
large_enum_variant (boxing one side of `Mixer` adds an indirection on the hot
path, and nothing gets optimised here without measuring).

The second of those is a debt, not a verdict, so it is now item 3 in
ROADMAP.md with the note that grouping the arguments needs a characterisation
test first: the 86 tests cover the lib, not the diagnostics in examples/.

Also the package description, still in Spanish.

Verified: clippy 0, build 0 warnings, 86 tests green.

+++ examples/coalitions.rs
@@ -55,7 +55,7 @@ fn main() {
         } else {
             bpb_with_skip(&model, &ds, n_train, n, &skips)
         };
-        if mask as usize == total - 1 {
+        if mask == total - 1 {
             base = bpb;
         }
         rows.push((mask, (mask as u32).count_ones(), bpb));
